Tweaks to make local dev zen work again
