+++ lib/lambda.js
@@ -131,6 +131,7 @@ async function withChromeProcess(opts, callback, { skipKill = false } = {}) {
   if (!manifest) throw new Error(`Missing manifest for ${opts.sessionId}`)
 
   const wrapper = new ChromeWrapper({
+    isRemote: true,
     headed: opts.headed,
     awsRegion: opts.awsRegion || process.env.AWS_REGION,
   })
