+++ lib/server.js
@@ -6,11 +6,15 @@ import Util from './util'
 import ChromeWrapper from './chrome_wrapper'
 import ChromeActions from './chrome_actions'
 
+// For debugging worker issues
+const HEADED_WORKERS = false
+
 class Server {
   constructor() {
+    const numWorkers = 4
     this.runId = 1
     this.head = null
-    this.workers = Array.construct(8, (id) => ({ id: id + 1 }))
+    this.workers = Array.construct(numWorkers, (id) => ({ id: id + 1 }))
     this.isLambda = false
     this.workerCount = 0
     this.workingSet = []
@@ -44,8 +48,8 @@ class Server {
 
     Zen.s3Sync.on('status', this.sendStatus.bind(this))
 
-    this.chrome = new ChromeWrapper() // headless chrome instance
-    this.chrome.launchLocal({ port: 9222 })
+    this.chrome = new ChromeWrapper({ headed: HEADED_WORKERS })
+    this.chrome.launchLocal({ port: 9222, headed: HEADED_WORKERS })
     new WebSocket.Server({ server }).on(
       'connection',
       this.onWebsocket.bind(this)
@@ -53,7 +57,7 @@ class Server {
 
     // create a server for each worker. This gives us different origins and isolates things like localStorage
     this.workersPromises = []
-    this.workers = Array.construct(8, (id) => {
+    this.workers = Array.construct(this.workers.length, (id) => {
       id = id + 1
       let port = Zen.config.port + id
       http.createServer(app).listen(port)
@@ -76,12 +80,10 @@ class Server {
     app.use(async (req, resp) => {
       const isRemote = req.url.match(/^\/worker/)
 
-      if (!isRemote) {
-        // Required to use WASM Sqlite.
-        // Currently wasm sqlite is only running on local dev.
-        resp.setHeader('Cross-Origin-Embedder-Policy', 'require-corp')
-        resp.setHeader('Cross-Origin-Opener-Policy', 'same-origin')
-      }
+      // Required to use WASM Sqlite.
+      // Currently wasm sqlite is only running on local dev.
+      resp.setHeader('Cross-Origin-Embedder-Policy', 'require-corp')
+      resp.setHeader('Cross-Origin-Opener-Policy', 'same-origin')
 
       resp.end(Zen.indexHtml(isRemote ? 'worker' : 'head'))
     })
@@ -110,7 +112,13 @@ class Server {
       }
 
       this.runId++
-      this.runOnLambda(msg)
+
+      if (this.workingSet.length < 25) {
+        this.runLocally()
+      } else {
+        this.runOnLambda(msg)
+      }
+
       this.sendStatus()
     }
   }
@@ -128,44 +136,35 @@ class Server {
     await Zen.s3Sync.run(Zen.indexHtml('worker', true))
     this.sendStatus()
 
-    // Needs the same pooling logic
-    runGroups.forEach(async (group) => {
-      const testNames = group.tests
-      try {
-        console.log('run:', testNames)
-        const response = await this.chromeActions.workTests({ testNames })
-        console.log('lambda response:', response)
-
-        if (startingRunId !== this.runId) return
-        this.onResults(response)
-      } catch (e) {
-        if (startingRunId !== this.runId) return
-        this.onResults(
-          testNames.map((fullName, testNumber) => {
-            return {
-              error: e.message,
-              fullName,
-              // batchId: workerId,
-              testNumber,
-            }
-          })
-        )
-      }
+    await Util.runTestsOnRemote({
+      zen: Zen,
+      opts: {
+        isLocal: true,
+        headed: false,
+        deflake: false,
+        maxAttempts: 3,
+      },
+      tests: this.workingSet,
+      onResult: (testResult, finalAttempt) => {
+        if (finalAttempt && startingRunId === this.runId) {
+          if (startingRunId !== this.runId) return
+          this.onResults([testResult])
+        }
+      },
     })
   }
 
   async runLocally() {
-    let startingRunId = this.runId,
-      remaining = this.workingSet.clone()
+    const startingRunId = this.runId
+    const remaining = this.workingSet.clone()
     this.isLambda = false
     this.workerCount = this.workers.length
+
     await Promise.all(this.workersPromises)
-    console.log(this.workers)
     this.workers.forEach(async (w) => {
       while (remaining.length > 0) {
-        console.log('sendTo Worker')
-        let result = await w.tab.setTest({ testName: remaining.pop() })
-        console.log('worker result', result)
+        const nextTest = remaining.pop()
+        let result = await w.tab.setTest({ testName: nextTest })
         if (!result || startingRunId !== this.runId) break // if the run was aborted
         this.onResults([result])
       }
