+++ lib/worker.js
@@ -28,6 +28,7 @@
 
   Zen.run = async function (opts) {
     runId = opts.runId
+
     let tests = Latte.flatten().filter((t) => t.fullName == opts.testName)
 
     if (opts.batch) {
